requirements.txt
anthropic
requests
beautifulsoup4
lxml
wordcloud
Pillow
